cheri: Apply the tasklet.cocci script
Run the tasklet.cocci script to fix tasklet calling API.

Coccinelle can do cross-file analysis provided that all the
affected files are in a single invocation. Thus to catch all
matches of the script use the following hack:
- Determine all directories that contain potential matches,
  e.g. by grepping for file that contain the string "tasklet".
- For each directory <DIR> run
	spatch -sp-file scripts/coccinelle/cheri/tasklet.cocci <DIR>/*.[ch]

+++ drivers/dma/switchtec_dma.c
@@ -820,7 +820,7 @@ unlock_and_exit:
 	return ret;
 }
 
-static void switchtec_dma_desc_task(unsigned long data)
+static void switchtec_dma_desc_task(uintptr_t data)
 {
 	struct switchtec_dma_chan *swdma_chan = (void *)data;
