+++ drivers/hsi/controllers/omap_ssi_core.c
@@ -208,7 +208,7 @@ static void ssi_gdd_complete(struct hsi_controller *ssi, unsigned int lch)
 	msg->actual_len = sg_dma_len(msg->sgt.sgl);
 }
 
-static void ssi_gdd_tasklet(unsigned long dev)
+static void ssi_gdd_tasklet(uintptr_t dev)
 {
 	struct hsi_controller *ssi = (struct hsi_controller *)dev;
 	struct omap_ssi_controller *omap_ssi = hsi_controller_drvdata(ssi);
